+++ src/providers/AuthProvider.jsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 import ResetPage from "../pages/passwordreset/[resetCode]";
 import AuthPage from "../screens/Auth";
 import LandingPage from "../components/LandingPage";
-import DayOfCheckin from "../pages/events/[eventId]/day-of-check-in";
+import DayOfCheckin from "../screens/DayOfCheckIn";
 import AddOrganizationModal from "../components/AddOrganizationModal";
 
 // AuthProvider wraps the entire application and makes sure only authenticated users can access the app
