+++ src/screens/DayOfCheckIn/helpers.js
